refactor model download

+++ enjoy/src/main/downloader.ts
@@ -2,7 +2,9 @@ import { ipcMain, app } from "electron";
 import path from "path";
 import fs from "fs";
 import mainWin from "@main/window";
+import log from "electron-log/main";
 
+const logger = log.scope("downloader");
 class Downloader {
   public tasks: Electron.DownloadItem[];
 
@@ -55,9 +57,11 @@ class Downloader {
           if (state === "completed") {
             resolve(item.getSavePath());
           } else {
-            fs.rmSync(item.getSavePath(), {
-              force: true,
-            });
+            if (fs.lstatSync(item.getSavePath()).isFile()) {
+              fs.rmSync(item.getSavePath(), {
+                force: true,
+              });
+            }
             resolve(undefined);
           }
         });
@@ -66,10 +70,14 @@ class Downloader {
   }
 
   cancel(filename: string) {
-    const task = this.tasks.find((t) => t.getFilename() === filename);
-    if (task && task.getState() === "progressing") {
-      task.cancel();
-    }
+    logger.debug("dashboard", this.dashboard());
+    this.tasks
+      .filter(
+        (t) => t.getFilename() === filename && t.getState() === "progressing"
+      )
+      .forEach((t) => {
+        t.cancel();
+      });
   }
 
   cancelAll() {
@@ -100,6 +108,7 @@ class Downloader {
       });
     });
     ipcMain.handle("download-cancel", (_event, filename) => {
+      logger.debug("download-cancel", filename);
       this.cancel(filename);
     });
     ipcMain.handle("download-cancel-all", () => {
